rpc: moveonly: untangle auxpow from rpc/mining.cpp
Aims to isolate auxpow code for readability and to make a clear
demarcation for this custom Dogecoin functionality. Only moves
code.

- Creates a header for mining/auxpow shared usage containing
  common functions between these functions.
- Moves all auxpow RPC methods into rpc/auxpow.cpp
